cleanup and reorganization
